Transient domain
