+++ packages/core/src/component.ts
@@ -15,7 +15,7 @@
 import type { IntlString, Plugin, StatusCode } from '@anticrm/platform'
 import { plugin } from '@anticrm/platform'
 import { Mixin, Version } from '.'
-import type { Account, AnyAttribute, ArrOf, AttachedDoc, Class, Collection, Doc, Interface, Obj, PluginConfiguration, PropertyType, Ref, RefTo, Space, Timestamp, Type } from './classes'
+import type { Account, AnyAttribute, ArrOf, AttachedDoc, Class, Collection, Doc, Interface, Obj, PluginConfiguration, PropertyType, Ref, RefTo, Space, Timestamp, Type, UserStatus } from './classes'
 import type { Tx, TxBulkWrite, TxCollectionCUD, TxCreateDoc, TxCUD, TxMixin, TxPutBag, TxRemoveDoc, TxUpdateDoc } from './tx'
 
 /**
@@ -55,7 +55,8 @@ export default plugin(coreId, {
     Collection: '' as Ref<Class<Collection<AttachedDoc>>>,
     Bag: '' as Ref<Class<Type<Record<string, PropertyType>>>>,
     Version: '' as Ref<Class<Version>>,
-    PluginConfiguration: '' as Ref<Class<PluginConfiguration>>
+    PluginConfiguration: '' as Ref<Class<PluginConfiguration>>,
+    UserStatus: '' as Ref<Class<UserStatus>>
   },
   space: {
     Tx: '' as Ref<Space>,
